fixed octave shift bug in sequencer (could go below zero and crash)

+++ MOUSE_DUAL/MOUSE_DUAL/src/sequencer.h
@@ -74,7 +74,7 @@ typedef struct _tSequencer
 	tNoteStack notestack;
 	uint8_t lengthCounter;
 	
-	int octave;
+	int8_t octave;
 	SequencerPatternType pattern;
 	int maxLength;
 	GlobalOptionType pitchOrTrigger;
